Ensure threshold and pubkeys size accomodate maximum for p2tr multisig(999)

+++ src/models/mod.rs
@@ -143,7 +143,7 @@ pub enum Requirement {
     /// Check signature from stack requirement (signature verified against a message)
     CheckSigFromStack { signature: String, pubkey: String, message: String },
     /// Check multisig requirement
-    CheckMultisig { signatures: Vec<String>, pubkeys: Vec<String>, threshold: u8 },
+    CheckMultisig { signatures: Vec<String>, pubkeys: Vec<String>, threshold: u16 },
     /// After requirement
     After { blocks: u64, timelock_var: Option<String> },
     /// Hash equal requirement
